# doc/apipost-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "\u62a5\u544a\u751f\u6210\u5e73\u53f0\u63a5\u53e3\u6587\u6863",
    "description": "ReportGenerate \u540e\u7aef\u5df2\u5b9e\u73b0\u63a5\u53e3\uff0c\u5305\u62ec\u7cfb\u7edf\u63a2\u6d3b\u3001\u6587\u4ef6\u7ba1\u7406\u3001\u6a21\u677f\u7ba1\u7406\u548c\u5065\u5eb7\u68c0\u67e5\u3002",
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "\u672c\u5730\u5f00\u53d1\u73af\u5883"
    }
  ],
...
  "paths": {
...
    "/api/files": {
      "post": {
        "tags": [
          "File"
        ],
        "summary": "\u4e0a\u4f20\u6587\u4ef6",
        "description": "\u4e0a\u4f20\u6587\u4ef6\u5230 MinIO\uff0c\u5e76\u5199\u5165 file_object \u6587\u4ef6\u5143\u6570\u636e\u3002\u53c2\u6570\u5fc5\u987b\u653e\u5728 Body \u7684 form-data \u4e2d\u3002",
        "operationId": "uploadFile",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestIdHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "bizType"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "\u4e0a\u4f20\u6587\u4ef6\uff1bApipost \u4e2d\u8bf7\u9009\u62e9 File \u7c7b\u578b\u3002"
                  },
                  "bizType": {
                    "type": "string",
                    "description": "\u6587\u4ef6\u4e1a\u52a1\u7c7b\u578b",
                    "enum": [
                      "TEMPLATE",
                      "REFERENCE",
                      "REPORT",
                      "PREVIEW",
                      "OTHER"
                    ],
                    "example": "REFERENCE"
                  },
                  "bizId": {
                    "type": "integer",
                    "format": "int64",
                    "nullable": true,
                    "description": "\u5173\u8054\u4e1a\u52a1 ID",
                    "example": 20001
                  },
                  "objectName": {
                    "type": "string",
                    "nullable": true,
                    "description": "MinIO \u5bf9\u8c61\u8def\u5f84\uff1b\u7b2c\u4e00\u6b21\u6d4b\u8bd5\u5efa\u8bae\u4e0d\u586b\u3002",
                    "example": "references/test/20001/demo.pdf"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "\u6210\u529f",
            "headers": {
              "X-Request-Id": {
                "description": "\u8bf7\u6c42\u94fe\u8def ID",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileObjectApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "\u8bf7\u6c42\u53c2\u6570\u9519\u8bef\u6216\u4e1a\u52a1\u6821\u9a8c\u5931\u8d25",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "\u670d\u52a1\u5185\u90e8\u9519\u8bef",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/{fileId}": {
      "get": {
        "tags": [
          "File"
        ],
        "summary": "\u67e5\u8be2\u6587\u4ef6\u8be6\u60c5",
        "description": "\u6839\u636e\u6587\u4ef6 ID \u67e5\u8be2\u6587\u4ef6\u5143\u6570\u636e\u3002fileId \u662f\u8def\u5f84\u53c2\u6570\uff0c\u586b\u5199\u5728 Path \u4e2d\uff0c\u4e0d\u653e Body\u3002",
        "operationId": "getFile",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestIdHeader"
          },
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "description": "\u6587\u4ef6 ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "\u6210\u529f",
            "headers": {
              "X-Request-Id": {
                "description": "\u8bf7\u6c42\u94fe\u8def ID",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileObjectApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "\u8bf7\u6c42\u53c2\u6570\u9519\u8bef\u6216\u4e1a\u52a1\u6821\u9a8c\u5931\u8d25",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "\u670d\u52a1\u5185\u90e8\u9519\u8bef",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "File"
        ],
        "summary": "\u903b\u8f91\u5220\u9664\u6587\u4ef6",
        "description": "\u903b\u8f91\u5220\u9664\u6587\u4ef6\u5143\u6570\u636e\u3002fileId \u662f\u8def\u5f84\u53c2\u6570\uff0c\u586b\u5199\u5728 Path \u4e2d\uff0c\u4e0d\u653e Body\u3002",
        "operationId": "deleteFile",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestIdHeader"
          },
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "description": "\u6587\u4ef6 ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "\u6210\u529f",
            "headers": {
              "X-Request-Id": {
                "description": "\u8bf7\u6c42\u94fe\u8def ID",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoidApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "\u8bf7\u6c42\u53c2\u6570\u9519\u8bef\u6216\u4e1a\u52a1\u6821\u9a8c\u5931\u8d25",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "\u670d\u52a1\u5185\u90e8\u9519\u8bef",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/{fileId}/access-url": {
      "get": {
        "tags": [
          "File"
        ],
        "summary": "\u83b7\u53d6\u6587\u4ef6\u4e34\u65f6\u8bbf\u95ee URL",
        "description": "\u751f\u6210\u6587\u4ef6\u4e34\u65f6\u4e0b\u8f7d\u6216\u9884\u89c8 URL\u3002fileId \u586b Path\uff1busage\u3001expireSeconds \u662f Query \u53c2\u6570\uff0c\u4e0d\u653e Body\u3002",
        "operationId": "getFileAccessUrl",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestIdHeader"
          },
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "description": "\u6587\u4ef6 ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "usage",
            "in": "query",
            "required": false,
            "description": "\u8bbf\u95ee\u7528\u9014",
            "schema": {
              "type": "string",
              "default": "DOWNLOAD",
              "enum": [
                "DOWNLOAD",
                "PREVIEW"
              ]
            },
            "example": "DOWNLOAD"
          },
          {
            "name": "expireSeconds",
            "in": "query",
            "required": false,
            "description": "\u4e34\u65f6 URL \u6709\u6548\u671f\uff0c\u6700\u5927 3600 \u79d2",
            "schema": {
              "type": "integer",
              "default": 600,
              "minimum": 1,
              "maximum": 3600
            },
            "example": 600
          }
        ],
        "responses": {
          "200": {
            "description": "\u6210\u529f",
            "headers": {
              "X-Request-Id": {
                "description": "\u8bf7\u6c42\u94fe\u8def ID",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileAccessUrlApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "\u8bf7\u6c42\u53c2\u6570\u9519\u8bef\u6216\u4e1a\u52a1\u6821\u9a8c\u5931\u8d25",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "\u670d\u52a1\u5185\u90e8\u9519\u8bef",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "XRequestIdHeader": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "description": "\u8bf7\u6c42\u94fe\u8def ID\uff1b\u4e0d\u4f20\u65f6\u540e\u7aef\u81ea\u52a8\u751f\u6210\u3002",
        "schema": {
          "type": "string"
        },
        "example": "REQ202606300001"
      }
    },
    "schemas": {
...
      "FileObjectResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10001
          },
          "fileCode": {
            "type": "string",
            "example": "F001"
          },
          "bizType": {
            "type": "string",
            "example": "REFERENCE"
          },
          "bizId": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "example": 20001
          },
          "originalFilename": {
            "type": "string",
            "example": "evaluation-material.pdf"
          },
          "objectName": {
            "type": "string",
            "example": "references/unknown/20001/F001/evaluation-material.pdf"
          },
          "bucketName": {
            "type": "string",
            "example": "report-generate"
          },
          "contentType": {
            "type": "string",
            "example": "application/pdf"
          },
          "fileExt": {
            "type": "string",
            "example": "pdf"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64",
            "example": 204800
          },
          "fileHash": {
            "type": "string",
            "nullable": true,
            "example": "sha256:xxxx"
          },
          "status": {
            "type": "string",
            "example": "ACTIVE"
          }
        }
      },
      "FileAccessUrlResponse": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "http://localhost:9000/report-generate/references/demo.pdf?X-Amz-Signature=xxx"
          },
          "expireSeconds": {
            "type": "integer",
            "example": 600
          },
          "usage": {
            "type": "string",
            "enum": [
              "DOWNLOAD",
              "PREVIEW"
            ],
            "example": "DOWNLOAD"
          },
          "expireAt": {
            "type": "string",
            "format": "date-time",
            "description": "\u4e34\u65f6 URL \u5931\u6548\u65f6\u95f4\uff0c\u540e\u7aef\u6309\u5f53\u524d\u65f6\u95f4\u52a0 expireSeconds \u8ba1\u7b97\u3002",
            "example": "2026-06-30T16:30:00+08:00"
          }
        }
      },
...
      "FileObjectApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/FileObjectResponse"
          },
          "requestId": {
            "type": "string",
            "example": "REQ202606300001"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "FileAccessUrlApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/FileAccessUrlResponse"
          },
          "requestId": {
            "type": "string",
            "example": "REQ202606300001"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "VoidApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "nullable": true
          },
          "requestId": {
            "type": "string",
            "example": "REQ202606300001"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "validation failed"
          },
          "data": {
            "nullable": true
          },
          "requestId": {
            "type": "string",
            "example": "REQ202606300001"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
    }
  }
}